Some fixes on post these will be changed
